split off spreadsheet functions from save_to_excel
